refactor(find-project): project-info component with instructions slot
